Cashu UX pass: - PaymentsHomeFragment header: when Cashu enabled, show sats-based fiat header text (leaves MOB MoneyView for now but displays Cashu fiat text) - PaymentsAddMoneyFragment already switches to mint quote/invoice via CashuMintQuoteUiHelper when Cashu is enabled - PaymentsValues: keep cashuEnabled flag; paymentsAvailability returns WITHDRAW_AND_SEND when Cashu is enabled}

+++ app/src/main/java/org/thoughtcrime/securesms/payments/engine/CashuEngine.kt
@@ -88,7 +88,7 @@ class CashuEngine(private val appContext: Context) : PaymentsEngine {
         amountSats = amountSats,
         feeSats = 0L,
         totalSats = amountSats,
-        expiresAtMs = cdkQuote.expiry,
+        expiresAtMs = cdkQuote.expiry.toLong(),
         invoiceBolt11 = cdkQuote.request,
         id = cdkQuote.id
       )
